Rollback rename of admin seed file
