+++ prisma/seed.ts
@@ -1,6 +1,6 @@
 import { sectors } from "./seeds/sectors";
 import {prisma} from "../lib/prisma";
-import AdminSeed from "./seeders/admin.seed";
+import AdminSeed from "./seeders/admin";
 
 async function main() {
   await prisma.sector.createMany({
